Handle devices that send headers and body separately.
If a device separates its headers and the body then we don't have an
easy way to identify that more data should be read. The easiest approach
may be to just override the read method and read more data from the
remote if higher layers indicate that we have an incomplete read.