Fix #61 amd close #67. Remove staging/gatt/*.
This switches us to use the upstream gatt dependency.

ble_impl/gatt.py contains our pending changes to the upstream library.
When it is merged this file can be removed.

ble_impl/device.py contains helpers we use for talking to HomeKit
devices (such as waiting for services to be resolved on connection).

This branch also removes a number of DeviceManager().run() that are no
longer required because of our enhanced Device() subclass.