Add BLE device tests
We implement a wrapper that maps Accessory models to an API that
looks like a BLE device accessed over our GATT library. This means
we implement both sides of the crypto and can 'round trip' (we are
sure that arent just writing gibberish to endpoints).

It's a fake rather than a mock. A mock would be more fragile to
changes in the implementation, but this does mean there is more
code required.