Serialization compat between ed25519 and py25519
ed25519 serializes its private key with the public key appended. Apart from
this the format is the same as py25119. So we trim the private key when
serializing, and append the public key when restoring a SigningKey. This should
mean the upgrade is transparent for existing pairing files and existing
accessories.