Basic threading functionality for 'get_events'
Added an optional semaphore that tells get_events to stop. Also, added a stop_events to cleanly tell the accessory to stop sending events to us.

Combined some of the functionality of get_events into put_characteristics, so put_characteristics is now more flexible (can be called with both the field that should be updated, as well as a default value to set multiple characteristics at once)