Add "auto" for pairing authentication method to BLE
Based on the feature flags, the best method for authentication is determined.
Note that no supported authentication method (both flags are not set) is
treated as "supports software authentication".
Having both flags not set may be the case for uncertified accessories, which
have neither co-processor nor the MFi authentication token.