Fix #223: Cache characteristics on BlePairing-initialization
If necessary, all services and characteristics will be cached on pairing
via BLE. This ensures that the pairing_data is in a valid state after
initializing the pairing object (which is important for e.g. BleSession).

Note that this behavior is similar to the Apple Home app, which also caches
all characteristics on pairing.